+++ src/io/aviso/rook/server.clj
@@ -3,10 +3,16 @@
   (:require [clojure.core.async :refer [alt! chan timeout >!]]
             [io.aviso.rook.async :as async]
             [io.aviso.rook.utils :as utils]
+            [io.aviso.binary :refer [format-binary]]
+            [io.aviso.toolchest.macros :refer [cond-let]]
             [io.aviso.toolchest.collections :refer [pretty-print]]
             [clojure.tools.logging :as l]
-            [ring.util.response :as r])
-  (:import (javax.servlet.http HttpServletResponse)))
+            [clojure.java.io :as io]
+            [ring.util.response :as r]
+            [medley.core :as medley]
+            [clojure.string :as str])
+  (:import [javax.servlet.http HttpServletResponse]
+           [java.io ByteArrayOutputStream InputStream]))
 
 (defn reloading-handler
   "Wraps a handler creator function such that the root handler is created fresh on each request;
@@ -23,15 +29,107 @@
     (l/info (utils/summarize-request request))
     (handler request)))
 
+(defn- read-bytes
+  [stream length]
+  (let [output-stream (ByteArrayOutputStream. (or length 2000))]
+    (io/copy stream output-stream)
+    (.toByteArray output-stream)))
+
+(defn- format-bytes
+  [label byte-array]
+  (let [length (if (some? byte-array)
+                 (alength byte-array)
+                 0)]
+    (if (zero? length)
+      ""
+      (format "%n%s (%,d bytes):%n%s"
+              label
+              length
+              (format-binary byte-array :ascii true :line-bytes 40)))))
+
+(defn- log-request-and-body [request ^bytes body-array]
+  (l/debugf "Request %s:%n%s%s"
+            (utils/summarize-request request)
+            (->> (dissoc request :uri :request-method :body)
+                 (medley/remove-vals nil?)
+                 pretty-print)
+            (format-bytes "Request Body" body-array)))
+
 (defn wrap-debug-request
-  "Logs a pretty-printed representation of each incoming request, at level debug."
+  "Logs a pretty-printed representation of each incoming request, at level debug. This includes output
+  of the body's content in ASCII and hex; this means that the entire body (if present) may be consumed
+  and stored in memory, and replaced with a new InputStream downstream (when debugging is enabled)."
   [handler]
   (fn [request]
-    (l/debugf "Request:%n%s" (pretty-print request))
-    (handler request)))
+    (cond-let
+      (not (l/enabled? :debug))
+      (handler request)
+
+      [{:keys [^InputStream body content-length]} request]
+
+      (nil? body)
+      (do
+        (log-request-and-body request nil)
+        (handler request))
+
+      (.markSupported body)
+      (let [_ (.mark body Integer/MAX_VALUE)
+            body-array (read-bytes body content-length)]
+        (log-request-and-body request body-array)
+        ;; Reset the stream back to its prior position before continuing.
+        (.reset body)
+        (handler request))
+
+      :else
+      (let [body-array (read-bytes body content-length)]
+        (log-request-and-body request body-array)
+        ;; We've consumed the :body of the request, must replace it with a new InputStream
+        ;; before continuing further.
+        (-> request
+            (assoc :body (io/input-stream body-array))
+            handler)))))
 
 (def ^:private not-found-response
   (utils/response HttpServletResponse/SC_NOT_FOUND))
+
+(defn- log-response
+  "Logs the response at debug level (if enabled) with nice formatting; returns a new response (since reading
+  a InputStream body is destructive)."
+  [response]
+  (cond-let
+    (not (l/enabled? :debug))
+    response
+
+    [body (:body response)]
+
+    (not (instance? InputStream body))
+    (do
+      (l/debugf "Response:%n%s" (pretty-print body))
+      response)
+
+    [mark-supported? (.markSupported body)
+     _ (and mark-supported? (.mark body Integer/MAX_VALUE))
+
+     content-length-str (get-in response [:headers "Content-Length"])
+
+     content-length (if (not (str/blank? content-length-str))
+                      (Integer/parseInt content-length-str))
+
+     body-array (read-bytes body content-length)
+
+     response' (dissoc response :body)]
+
+    :else
+    (do
+      (l/debugf "Response:%n%s%s"
+                (pretty-print response')
+                (format-bytes "Response Body" body-array))
+      ;; Mark is more commonly supported on the response side than on the request side.
+      (if mark-supported?
+        (do
+          (.reset body)
+          response))
+      (assoc response :body (io/input-stream body-array)))))
 
 (defn wrap-with-timeout
   "Asynchronous Ring middleware that applies a timeout to a request. When the timeout occurs, a 504 response is returned
@@ -56,9 +154,7 @@
                                     response)
                        handler-ch ([response]
                                     (if response
-                                      (do
-                                        (l/debugf "Response:%n%s" (pretty-print response))
-                                        response)
+                                      (log-response response)
                                       (do
                                         (l/debugf "Handler for %s closed response channel." (utils/summarize-request request))
                                         not-found-response))))))))